test: 🧪 move tests for MarkOptional to separate file
